pyproject.toml
[tool.poetry]
name = "brickbreaker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.scripts]
main = "brickbreaker.main:main"

[tool.poetry.dependencies]
python = "^3.9"
autopep8 = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
PyOpenGL = "^3.1.6"
glfw = "^2.5.3"
pyrr = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
